Final_netinsight/requirements.txt
Flask>=2.2.0
SQLAlchemy>=1.4.0
textfsm>=1.1.0
requests>=2.28.0
python-dateutil>=2.8.0
iperf3
dnspython>=2.2.0
speedtest-cli>=2.1.3
